requirements.txt
numpy
opencv-python
Pillow
matplotlib

pyautogui
scipy